backend/requirements.txt
flask==1.1.1
flask-cors==3.0.8
mysql-connector==2.2.9
